Windows: Compile containerd.exe (non-functional at this point)
