+++ cmd/containerd/builtins_windows.go
@@ -1,0 +1,8 @@
+package main
+
+// register containerd builtins here
+
+//	_ "github.com/containerd/containerd/services/content"
+//	_ "github.com/containerd/containerd/services/execution"
+//	_ "github.com/containerd/containerd/services/healthcheck"
+//	_ "github.com/containerd/containerd/services/metrics"
