+++ content/store_windows.go
@@ -1,0 +1,10 @@
+package content
+
+import (
+	"os"
+	"time"
+)
+
+func getStartedAt(fi os.FileInfo) time.Time {
+	return fi.ModTime()
+}
